Created the npm plugin which can be installed by anyone
