+++ client/package.json
@@ -9,6 +9,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "avsar-thirdparty-plugin": "^1.0.1",
     "credential-handler-polyfill": "^3.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
